Heartbleed auto script
With this script you can type the IP of a server and check if it is
vulnerable to heartbleed. Also you set the number of attacks against the
server. If the server is vulnerable it stores the private
key,username,passwords and emails to a txt file. You must have
metasploit in order to run the script.